add LegacyAggregateBasedEventStorageEngineUtils
